Split SerialiseBech32 module out of Cardano.Api.Typed
